% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMidPoints.R
\name{calcMidPoints}
\alias{calcMidPoints}
\title{Calculate midpoints of a vector}
\usage{
calcMidPoints(v)
}
\arguments{
\item{v}{- a numeric vector}
}
\value{
a numeric vector with 1 less element than v with corresponding midpoints of intervals
}
\description{
Function to calculate the midpoints of a vector.
}
