% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsUtilities.R
\name{createAggregatorMap}
\alias{createAggregatorMap}
\title{Create a dims aggregator map}
\usage{
createAggregatorMap(dms_frm, dms_to, keepOrigDims = FALSE)
}
\arguments{
\item{dms_frm}{- aggregating from dimensions map}

\item{dms_to}{- aggregating to dimensions map}

\item{keepOrigDims}{- flag (T/F) to keep original dimensions (T) or aggregated dimensions (F)}
}
\value{
map from original dimensions to aggregated dimensions.
   * column "idx_to" gives index into vector with aggregated dimensions
   * column "idx_from" gives index into vector with original dimensions
}
\description{
Function to create an aggregator map from one set of dimensions to another.
}
