% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{Create model}
\usage{
createModel(inputs, mode = "debug")
}
\arguments{
\item{inputs}{- list of inputs to the model}

\item{mode}{- run mode ("run","debug")}
}
\description{
Function to create a gmacs model
}
\details{
Creates a gmacs model with provided inputs.
#Example
inputs = list();
inputs$data = list(dims=1:3);
inputs$pars = list(pars=4:7);
model = tmbGMACS::createModel(inputs,mode="debug");
}
