% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_other_functions.R
\name{.onLoad}
\alias{.onLoad}
\title{Allow Quarto-based vignettes}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{- library name (?)}

\item{pkgname}{- package name}
}
\value{
see [tools::vignetteEngine()].
}
\description{
Function to allow Quarto-based vignettes
}
\details{
Uses [tools::vignetteEngine()] to create a Quarto-based vignette engine
for this package.
}
