% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsUtilities.R
\name{dropDims}
\alias{dropDims}
\title{Drop selected dimensions from a dimensions map}
\usage{
dropDims(dms, drop)
}
\arguments{
\item{dms}{- dimensions map}

\item{drop}{- vector (character or integer) with dimensions to drop}
}
\value{
dimensions map without dropped dimensions
}
\description{
Function to drop selected dimensions from a dimensions map.
}
\details{
If \code{drop} is an integer vector, the values
specify the order of dimensions to drop.
}
