% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsFunctions.R
\name{getFullDimsIndex}
\alias{getFullDimsIndex}
\title{Get the 1-d index associated with a vector of dimension indices for a fully-crossed expansion}
\usage{
getFullDimsIndex(v, n, rev = FALSE)
}
\arguments{
\item{v}{- vector of dimension indices}

\item{n}{- vector of the number of corresponding dimension levels}

\item{rev}{- flag to get the index for the reverse order}
}
\value{
an integer, the 1-d index into the fully-crossed dimensions array
}
\description{
Function to get the 1-d index associated with a vector of dimension indices
for a fully-crossed expansion of the dimensions.
}
\details{
If \code{rev} is false (the default), the index is calculated with the
first dimension cycling the fastest; otherwise, the index is calculated with the
last dimension cycling the fastest.
}
\examples{
#--examples for 3 dimensions, each with 5 levels
require(tmbGMACS);
getFullDimsIndex(c(1,1,1),c(5,5,5));
getFullDimsIndex(c(5,1,1),c(5,5,5));
getFullDimsIndex(c(1,5,1),c(5,5,5));
getFullDimsIndex(c(1,1,5),c(5,5,5));
getFullDimsIndex(c(5,5,1),c(5,5,5));
getFullDimsIndex(c(1,5,5),c(5,5,5));
getFullDimsIndex(c(5,5,5),c(5,5,5));
getFullDimsIndex(c(1,1,1),c(5,5,5),TRUE);
getFullDimsIndex(c(5,1,1),c(5,5,5),TRUE);
getFullDimsIndex(c(1,5,1),c(5,5,5),TRUE);
getFullDimsIndex(c(1,1,5),c(5,5,5),TRUE);
getFullDimsIndex(c(5,5,1),c(5,5,5),TRUE);
getFullDimsIndex(c(1,5,5),c(5,5,5),TRUE);
getFullDimsIndex(c(5,5,5),c(5,5,5),TRUE);
}
