% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionsUtilities.R
\name{keepDims}
\alias{keepDims}
\title{Keep only selected dimensions from a dimensions map}
\usage{
keepDims(dms, keep)
}
\arguments{
\item{dms}{- dimensions map}

\item{drop}{- vector (character or integer) with dimensions to keep}
}
\value{
dimensions map with only kept dimensions
}
\description{
Function to keep only selected dimensions from a dimensions map.
}
\details{
If \code{keep} is an integer vector, the values
specify the order of dimensions to keep.
}
