% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_other_functions.R
\name{tibbleRowToString}
\alias{tibbleRowToString}
\title{Convert a tibble row to a string}
\usage{
tibbleRowToString(tbl_rw)
}
\arguments{
\item{tbl_rw}{- a row of a tibble}
}
\value{
a string
}
\description{
Function to convert a tibble row to a string.
}
