% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbGMACS-package.R
\docType{package}
\name{tmbGMACS-package}
\alias{tmbGMACS}
\alias{tmbGMACS-package}
\title{tmbGMACS: A generalized stock assessment modeling framework for crustacean fishery stocks}
\description{
TMB implementation of the GMACS stock assessment model. This package provides a modeling framework in R using TMB based on the GMACS stock assessment framework developed in ADMB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wStockhausen/tmbGMACS}
  \item Report bugs at \url{https://github.com/wStockhausen/tmbGMACS/issues}
}

}
\keyword{internal}
